qcheck: remove alignment check for compressed pointers
Sorry, I misread the spec on the last commit. Compressed
pointers do not have to align to /anything/ at all.

Corruption detection for compressed images remains woefully broken.
Unfortunately, this also extends to misclassifying images with corruption
that appears to look like valid compressed pointers.

Corrupt images with pointers that manage to trick the existing
validation may appear to have corrupted pointers that occur later
in the file that appear to fail the metadata overlap check,
when in reality the compressed pointer was the problem.

So qcheck currently misidentifies the location of which pointers
are the problem when dealing with evidently compressed pointers.